Refactor likes and comments in API
